add CapabilityChecker and CorrectnessChecker
