+++ core/src/types/operator/builder.rs
@@ -472,7 +472,7 @@ impl<A: Access> OperatorBuilder<A> {
         OperatorBuilder { accessor }
             .layer(ErrorContextLayer)
             .layer(CompleteLayer)
-            .layer(CapabilityCheckLayer::default())
+            .layer(CorrectnessChecker)
     }
 
     /// Create a new layer with static dispatch.
